Separate core types
