+++ decidim-api/app/types/decidim/api/localized_string_type.rb
@@ -3,7 +3,7 @@ module Decidim
   module Api
     # This type represents a localized string in a single language.
     LocalizedStringType = GraphQL::ObjectType.define do
-      name "LocalizedStringFieldType"
+      name "LocalizedStringField"
       description "Represents a particular translation of a LocalizedStringType"
 
       field :locale, !types.String, "The standard locale of this translation."
