+++ decidim-api/app/types/decidim/api/query_type.rb
@@ -5,15 +5,6 @@ module Decidim
     QueryType = GraphQL::ObjectType.define do
       name "Query"
       description "The root query of this schema"
-
-      field :processes do
-        type !types[ProcessType]
-        description "Lists all processes."
-
-        resolve ->(_obj, _args, ctx) {
-          ctx[:current_organization].participatory_processes
-        }
-      end
     end
   end
 end
