+++ decidim-api/decidim-api.gemspec
@@ -14,7 +14,6 @@ Gem::Specification.new do |s|
 
   s.files = Dir["{app,config,db,lib}/**/*", "LICENSE.txt", "Rakefile", "README.md"]
 
-  s.add_dependency "decidim-core", Decidim.version
   s.add_dependency "rails", *Decidim.rails_version
   s.add_dependency "graphql", "~> 1.2.5"
   s.add_dependency "graphiql-rails", "~> 1.4.0"
