+++ decidim-api/spec/types/query_type_spec.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 require "spec_helper"
+require "decidim/api/test/type_helpers"
 
 module Decidim
   module Api
     describe QueryType do
-      include TypeHelpers
+      include Test::TypeHelpers
 
       describe "processes" do
         let!(:process1) { create(:participatory_process, organization: current_organization) }
