+++ decidim-api/spec/types/translated_field_type_spec.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 require "spec_helper"
+require "decidim/api/test/type_helpers"
 
 module Decidim
   module Api
     describe TranslatedFieldType do
-      include TypeHelpers
+      include Test::TypeHelpers
 
       let(:model) do
         {
