+++ decidim-core/app/types/decidim/process_step_type.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+module Decidim
+  # This type represents a step on a participatory process.
+  ProcessStepType = GraphQL::ObjectType.define do
+    name "ProcessStep"
+    description "A participatory process step"
+
+    field :id, !types.ID, "The unique ID of this step."
+
+    field :process do
+      type ProcessType
+      description "The participatory process in which this step belongs to."
+      property :participatory_process
+    end
+
+    field :title, Api::TranslatedFieldType, "The title of this step"
+
+    field :shortDescription do
+      type Api::TranslatedFieldType
+      description "A short description of the step."
+      property :short_description
+    end
+  end
+end
