+++ decidim-core/app/types/decidim/process_type.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+module Decidim
+  # This type represents a ParticipatoryProcess.
+  ProcessType = GraphQL::ObjectType.define do
+    name "Process"
+    description "A participatory process"
+
+    field :id, !types.ID, "The Process' unique ID"
+
+    field :title, Api::TranslatedFieldType, "The title of this process."
+
+    connection :steps, ProcessStepType.connection_type do
+      description "All the steps of this process."
+    end
+  end
+end
